requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0
python-dateutil==2.8.2
flask-cors==3.0.10
